lambda/pyproject.toml
[tool.poetry]
name = "weather-checker"
version = "0.1.0"
description = ""
authors = ["mozkzki <mozkzki@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
selenium = "3.141.0"
Pillow = "^8.3.2"
requests = "2.25.1"
moz-image = {git = "https://github.com/mozkzki/moz-image.git", rev = "master"}

[tool.poetry.dev-dependencies]
flake8 = "4.0.1"
mypy = "0.910"
pytest = "6.2.5"
black = "21.9b0"
pytest-cov = "3.0.0"
coverage = "6.0.2"
isort = "5.9.3"
regex = "2021.9.30"
types-requests = "^2.25.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
